Clear the lints CI fails the build on
`cargo clippy --all -- -D warnings` has been failing. It stopped at the first
crates that would not compile, so the 27 errors it printed were only what it
reached: clearing those let the crates behind them build and report their own,
five rounds deep, across arm32_cpu, wie_backend, wie_core_arm, wie_jvm_support,
wie_wipi_c, wie_wipi_java, wie_ktf, wie_lgt and wie_android.

Every fix here is the lint's own rewrite and keeps what the code did:
`x + 1 <= y` becomes `x < y` without the add that would wrap, index loops
become iterators, nested `if`s become let-chains, `a * b >> c` gains the
parentheses it already parsed with, and manual `Default`/`matches!`/`contains`
/`clamp` shapes become the real ones.

Two judgements worth naming. Code that is deliberately unreached is kept and
explained rather than deleted: the JIT frontend's `FastOp` variants,
`ends_trace` and `pages_base`, and the LGT time API, are implemented but not
wired to a dispatch table yet. And the twelve argument-count lints in
`org.kwis.msp.lwc` are allowed once at the module, because every method there
is a JVM bridge whose parameters are the Java method's own plus the three the
bridge always carries.

`cargo fmt --all --check`, `cargo clippy --all -- -D warnings`, the same for
wasm32-unknown-unknown, and `cargo test --workspace` all pass.

+++ wie_android/src/filesystem.rs
@@ -247,7 +247,7 @@ impl Filesystem for AndroidFilesystem {
             return None;
         }
 
-        Some((stats.f_bsize as u64).saturating_mul(stats.f_blocks as u64))
+        Some((stats.f_bsize as u64).saturating_mul(stats.f_blocks))
     }
 
     async fn available_space(&self, aid: &str) -> Option<u64> {
@@ -275,7 +275,7 @@ impl Filesystem for AndroidFilesystem {
             return None;
         }
 
-        Some((stats.f_bsize as u64).saturating_mul(stats.f_bavail as u64))
+        Some((stats.f_bsize as u64).saturating_mul(stats.f_bavail))
     }
 
     async fn list(&self, aid: &str, path: &str) -> Option<Vec<String>> {
